requirements.txt
fastapi==0.75.0
uvicorn==0.17.0
openai==0.11.3
requests==2.26.0
pyttsx3==2.90
whisper==1.0
beautifulsoup4==4.10.0
faiss-cpu==1.7.2
streamlit==1.11.0
